Refactored data layer to conform with other layers
