+++ LudoCompose/app/src/main/java/hu/bme/aut/android/ludocompose/data/model/GameWithPlayers.kt
@@ -14,18 +14,19 @@
  * limitations under the License.
  */
 
-package hu.bme.aut.android.ludocompose.data.pojos
+package hu.bme.aut.android.ludocompose.data.model
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import hu.bme.aut.android.ludocompose.data.entities.PlayerEntity
-import hu.bme.aut.android.ludocompose.data.entities.TokenEntity
+import androidx.room.TypeConverters
+import hu.bme.aut.android.ludocompose.data.converters.LocalDateConverter
 
-data class PlayerWithTokens(
-    @Embedded val player: PlayerEntity,
+@TypeConverters(LocalDateConverter::class)
+data class GameWithPlayers(
+    @Embedded val game: GameEntity,
     @Relation(
         parentColumn = "id",
-        entityColumn = "player_id",
-        entity = TokenEntity::class,
-    ) val tokens: List<TokenEntity>
+        entityColumn = "game_id",
+        entity = PlayerEntity::class,
+    ) val players: List<PlayerWithTokens>
 )
