--- LudoCompose/app/src/main/java/hu/bme/aut/android/ludocompose/data/pojos/GameWithPlayers.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright © 2023 Benjamin Kis
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package hu.bme.aut.android.ludocompose.data.pojos
-
-import androidx.room.Embedded
-import androidx.room.Relation
-import androidx.room.TypeConverters
-import hu.bme.aut.android.ludocompose.data.converters.LocalDateConverter
-import hu.bme.aut.android.ludocompose.data.entities.GameEntity
-import hu.bme.aut.android.ludocompose.data.entities.PlayerEntity
-
-@TypeConverters(LocalDateConverter::class)
-data class GameWithPlayers(
-    @Embedded val game: GameEntity,
-    @Relation(
-        parentColumn = "id",
-        entityColumn = "game_id",
-        entity = PlayerEntity::class,
-    ) val players: List<PlayerWithTokens>
-)
